fixes volume synchronization, consolidates code
and some beautification
