+++ microscopynodes/handle_blender_structs/__init__.py
@@ -3,3 +3,4 @@ from .collection_handling import *
 from .node_handling import *
 from .progress_handling import *
 from .props import *
+from .array_handling import *
