+++ microscopynodes/handle_blender_structs/props.py
@@ -27,7 +27,7 @@ bpy.types.Scene.MiN_remake = bpy.props.BoolProperty(
 
 bpy.types.Scene.MiN_load_start_frame = bpy.props.IntProperty(
     name = "", 
-    description = "Force remaking vdb files",
+    description = "First timeframe to be loaded",
     default = 0,
     min=0,
     soft_max=10000,
@@ -35,7 +35,7 @@ bpy.types.Scene.MiN_load_start_frame = bpy.props.IntProperty(
 
 bpy.types.Scene.MiN_load_end_frame = bpy.props.IntProperty(
     name = "", 
-    description = "Force remaking vdb files",
+    description = "Last timeframe to be loaded.",
     default = 100,
     soft_max= 10000,
     min=0,
